+++ docs/fr_FR/changelog.md
@@ -3,6 +3,8 @@
 >**IMPORTANT**
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
+# 29/11/2023
+- ajout des commandes setClosureAndLinearSpeed, setPositionAndLinearSpeed et setPosition
 
 # 19/10/2023